frondend

+++ backend/src/main/java/_blog/com/_blog/services/PostServ.java
@@ -13,6 +13,7 @@ import _blog.com._blog.dto.PostConvert;
 import _blog.com._blog.repositories.CommentsRepositories;
 import _blog.com._blog.repositories.PostRepositery;
 import _blog.com._blog.repositories.ReactionRepo;
+import _blog.com._blog.repositories.ReportRepostiry;
 import _blog.com._blog.utils.PostReq;
 import _blog.com._blog.utils.Upload;
 
@@ -22,13 +23,15 @@ public class PostServ {
     private final NotifacationSer notifacationSer;
     private final ReactionRepo reactionRepo;
     private final CommentsRepositories commentsRepositories;
+    private final ReportRepostiry reportRepostiry;
 
     public PostServ(PostRepositery PostRepositery, NotifacationSer notifacationSer, ReactionRepo reactionRepo,
-            CommentsRepositories commentsRepositories) {
+            CommentsRepositories commentsRepositories , ReportRepostiry reportRepostiry) {
         this.postRepositery = PostRepositery;
         this.notifacationSer = notifacationSer;
         this.reactionRepo = reactionRepo;
         this.commentsRepositories = commentsRepositories;
+        this.reportRepostiry = reportRepostiry;
     }
 
     @Transactional
@@ -56,10 +59,11 @@ public class PostServ {
         Post Post = postRepositery.findById(Postid)
                 .orElseThrow(() -> new ProgramExeption(400, "Post not found"));
 
-        if (Post.getUser().getId() == user.getId() || user.getRole() == "ADMIN") {
+        if ((Post.getUser().getId() == user.getId() && !Post.isHide()) || user.getRole() == "ADMIN") {
             reactionRepo.deleteByPost(Postid);
             notifacationSer.deleteNotifactionByPostid(Postid);
             commentsRepositories.deleteByPostid(Postid);
+            reportRepostiry.deleteByPost(Postid);
             postRepositery.deleteById(Postid);
         } else {
             throw new ProgramExeption(400, "you can't delet this post");
